:D now kind of cache!
ok, a little improvement.
now all the items of averypage loaded are stored in-memory on a
different level of an array (I'd just added multilevel to the array of
items)
now as a result the pages loaded by second time (like going to second
page and then back to first page) are loaded faster because is skipped
the ajax call.
it only happens if the search query remains the same.